move listnetworks to user api

# Conflicts:
#	clients/remote/src/dapp-api/server.test.ts
