+++ api-specs/openrpc-dapp-api.json
@@ -142,24 +142,6 @@
                 }
             },
             "description": "Proxy for the JSON-API endpoints. Injects authorization headers automatically."
-        },
-        {
-            "name": "listNetworks",
-            "params": [],
-            "result": {
-                "name": "result",
-                "schema": {
-                    "title": "ListNetworksResult",
-                    "type": "object",
-                    "properties": {
-                        "networks": {
-                            "type": "array",
-                            "$ref": "#/components/schemas/Network"
-                        }
-                    },
-                    "required": ["networks"]
-                }
-            }
         }
     ],
     "components": {
@@ -179,46 +161,6 @@
                     }
                 },
                 "required": ["response"]
-            },
-            "Network": {
-                "title": "Network",
-                "type": "object",
-                "description": "Structure representing the connected Networks",
-                "properties": {
-                    "name": {
-                        "type": "string",
-                        "description": "Name of network"
-                    },
-                    "description": {
-                        "type": "string",
-                        "description": "Description of network"
-                    }
-                }
-            },
-            "Auth": {
-                "title": "type",
-                "type": "object",
-                "description": "Represents the type of auth (implicit or password) for a specified network",
-                "properties": {
-                    "type": {
-                        "type": "string"
-                    },
-                    "tokenUrl": {
-                        "type": "string"
-                    },
-                    "scope": {
-                        "type": "string"
-                    },
-                    "clientId": {
-                        "type": "string"
-                    },
-                    "domain": {
-                        "type": "string"
-                    },
-                    "audience": {
-                        "type": "string"
-                    }
-                }
             },
             "UserUrl": {
                 "title": "UserUrl",
