+++ clients/remote/src/dapp-api/controller.ts
@@ -21,6 +21,4 @@ export const dappController = (store: Store) =>
         prepareExecute: async (params: PrepareExecuteParams) => null,
         prepareReturn: async (params: PrepareReturnParams) =>
             Promise.resolve({}),
-        listNetworks: async () =>
-            Promise.resolve({ networks: await store.listNetworks() }),
     })
