+++ clients/remote/src/dapp-api/rpc-gen/index.ts
@@ -5,7 +5,6 @@ import { DarsAvailable } from './typings.js'
 import { PrepareReturn } from './typings.js'
 import { PrepareExecute } from './typings.js'
 import { LedgerApi } from './typings.js'
-import { ListNetworks } from './typings.js'
 
 export type Methods = {
     connect: Connect
@@ -13,7 +12,6 @@ export type Methods = {
     prepareReturn: PrepareReturn
     prepareExecute: PrepareExecute
     ledgerApi: LedgerApi
-    listNetworks: ListNetworks
 }
 
 function buildController(methods: Methods) {
@@ -23,7 +21,6 @@ function buildController(methods: Methods) {
         prepareReturn: methods.prepareReturn,
         prepareExecute: methods.prepareExecute,
         ledgerApi: methods.ledgerApi,
-        listNetworks: methods.listNetworks,
     }
 }
