+++ clients/remote/src/dapp-api/rpc-gen/typings.ts
@@ -46,24 +46,6 @@ export interface JsPrepareSubmissionResponse {
     preparedTransactionHash?: StringZK0Xb1WV
     [k: string]: any
 }
-/**
- *
- * Name of network
- *
- */
-export type StringWupREggx = string
-/**
- *
- * Description of network
- *
- */
-export type StringM72S2Xxb = string
-/**
- *
- * Structure representing the connected Networks
- *
- */
-export type Network = any[]
 export interface PrepareReturnParams {
     commands: JsCommands
     [k: string]: any
@@ -98,16 +80,12 @@ export interface LedgerApiResult {
     response: StringDoaGddGA
     [k: string]: any
 }
-export interface ListNetworksResult {
-    networks: Network
-    [k: string]: any
-}
 /**
  *
  * Generated! Represents an alias to any of the provided schemas
  *
  */
-export type AnyOfPrepareReturnParamsPrepareExecuteParamsLedgerApiParamsConnectResultDarsAvailableResultPrepareReturnResultPrepareExecuteResultLedgerApiResultListNetworksResult =
+export type AnyOfPrepareReturnParamsPrepareExecuteParamsLedgerApiParamsConnectResultDarsAvailableResultPrepareReturnResultPrepareExecuteResultLedgerApiResult =
 
         | PrepareReturnParams
         | PrepareExecuteParams
@@ -117,7 +95,6 @@ export type AnyOfPrepareReturnParamsPrepareExecuteParamsLedgerApiParamsConnectRe
         | PrepareReturnResult
         | PrepareExecuteResult
         | LedgerApiResult
-        | ListNetworksResult
 export type Connect = () => Promise<ConnectResult>
 export type DarsAvailable = () => Promise<DarsAvailableResult>
 export type PrepareReturn = (
@@ -127,4 +104,3 @@ export type PrepareExecute = (
     params: PrepareExecuteParams
 ) => Promise<PrepareExecuteResult>
 export type LedgerApi = (params: LedgerApiParams) => Promise<LedgerApiResult>
-export type ListNetworks = () => Promise<ListNetworksResult>
