+++ clients/remote/src/dapp-api/server.test.ts
@@ -41,44 +41,3 @@ test('call connect rpc', async () => {
         },
     })
 })
-
-test('call connect rpc', async () => {
-    const response = await request(dapp(authService, store))
-        .post('/rpc')
-        .send({ jsonrpc: '2.0', id: 0, method: 'listNetworks', params: [] })
-        .set('Accept', 'application/json')
-
-    expect(response.body).toEqual({
-        jsonrpc: '2.0',
-        id: 0,
-        result: {
-            networks: [
-                {
-                    name: 'xyz',
-                    description: 'name1',
-                    ledgerApi: { baseUrl: 'https://test' },
-                    auth: {
-                        type: 'password',
-                        tokenUrl: 'tokenUrl',
-                        grantType: 'password',
-                        scope: 'openid',
-                        clientId: 'wk-service-account',
-                    },
-                },
-                {
-                    name: 'abc',
-                    description: 'dex idp',
-                    ledgerApi: { baseUrl: 'https://test' },
-                    auth: {
-                        type: 'implicit',
-                        domain: 'dex.com',
-                        audience: 'https://daml.com/jwt/aud/participant/wk-app',
-                        scope: 'openid',
-                        clientId: 'wk-service-account2',
-                    },
-                },
-            ],
-        },
-    })
-    expect(response.statusCode).toBe(200)
-})
